expenseTracker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
astroid==3.3.9
async-timeout==5.0.1
attrs==25.3.0
awsebcli==3.21.0
blessed==1.20.0
boto3==1.35.57
botocore==1.35.63
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
fonttools==4.56.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
isort==6.0.1
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.4
mccabe==0.7.0
multidict==6.2.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.10.1
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.0
PyJWT==2.10.1
pylint==3.3.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tomli==2.2.1
tomlkit==0.13.2
twilio==9.5.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
yarl==1.18.3
zipp==3.21.0
